chore: fix smoke tests to account for new release versions within a workspace (#8143)
A reworking of the smoke tests to accommodate a recent PR where the
check to see if the version has already been published conflicted with
the smoke tests. A fix was implemented to add prerelease tags to address
this issue. However, when the version bumps are legitimate and a
prerelease tag is added, the workspace dependencies still have the
version prior to the prerelease tag. This PR adds the prerelease tag and
bumps all versions used in the workspace.